requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-slugify>=8.0.0
python-multipart>=0.0.9
aiofiles>=23.2.1
